+++ packages/atproto/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Supported `com.atproto.server.describeServer` as `findServerInfo` in `ServersService`. ([#260](https://github.com/myConsciousness/atproto.dart/issues/260))
 - Supported `com.atproto.server.createInviteCodes` as `createInviteCodes` in `ServersService`. ([#258](https://github.com/myConsciousness/atproto.dart/issues/258))
+- `identifier` in `createSession` is now required.
 
 ## v0.5.0